+++ lib/router.js
@@ -54,12 +54,12 @@ router.get('/weibo/search/hot', lazyloadRouteHandler('./routes/weibo/search/hot'
 router.get('/weibo/super_index/:id/:type?/:routeParams?', lazyloadRouteHandler('./routes/weibo/super_index'));
 router.get('/weibo/oasis/user/:userid', lazyloadRouteHandler('./routes/weibo/oasis/user'));
 
-// 贴吧
-router.get('/tieba/forum/:kw', lazyloadRouteHandler('./routes/tieba/forum'));
-router.get('/tieba/forum/good/:kw/:cid?', lazyloadRouteHandler('./routes/tieba/forum'));
-router.get('/tieba/post/:id', lazyloadRouteHandler('./routes/tieba/post'));
-router.get('/tieba/post/lz/:id', lazyloadRouteHandler('./routes/tieba/post'));
-router.get('/tieba/user/:uid', lazyloadRouteHandler('./routes/tieba/user'));
+// 贴吧 migrated to v2
+// router.get('/tieba/forum/:kw', lazyloadRouteHandler('./routes/tieba/forum'));
+// router.get('/tieba/forum/good/:kw/:cid?', lazyloadRouteHandler('./routes/tieba/forum'));
+// router.get('/tieba/post/:id', lazyloadRouteHandler('./routes/tieba/post'));
+// router.get('/tieba/post/lz/:id', lazyloadRouteHandler('./routes/tieba/post'));
+// router.get('/tieba/user/:uid', lazyloadRouteHandler('./routes/tieba/user'));
 
 // 网易云音乐
 router.get('/ncm/playlist/:id', lazyloadRouteHandler('./routes/ncm/playlist'));
