+++ lib/v2/baidu/maintainer.js
@@ -1,0 +1,7 @@
+module.exports = {
+    '/tieba/forum/:kw': ['u3u'],
+    '/tieba/forum/good/:kw/:cid?': ['u3u'],
+    '/tieba/post/:id': ['u3u'],
+    '/tieba/post/lz/:id': ['u3u'],
+    '/tieba/user/:uid': ['igxlin', 'nczitzk'],
+};
