+++ lib/v2/baidu/router.js
@@ -1,0 +1,7 @@
+module.exports = (router) => {
+    router.get('/tieba/forum/:kw', require('./tieba/forum'));
+    router.get('/tieba/forum/good/:kw/:cid?', require('./tieba/forum'));
+    router.get('/tieba/post/:id', require('./tieba/post'));
+    router.get('/tieba/post/lz/:id', require('./tieba/post'));
+    router.get('/tieba/user/:uid', require('./tieba/user'));
+};
