+++ lib/v2/baidu/tieba/post.js
@@ -1,5 +1,7 @@
 const cheerio = require('cheerio');
 const got = require('@/utils/got');
+const { parseDate } = require('@/utils/parse-date');
+const timezone = require('@/utils/timezone');
 
 /**
  * 获取最新的帖子回复（倒序查看）
@@ -11,15 +13,13 @@ const got = require('@/utils/got');
  * @returns
  */
 async function getPost(id, lz = 0, pn = 7e6) {
-    const { data } = await got({
-        method: 'get',
-        url: `https://tieba.baidu.com/p/${id}?see_lz=${lz}&pn=${pn}&ajax=1`,
+    const { data } = await got(`https://tieba.baidu.com/p/${id}?see_lz=${lz}&pn=${pn}&ajax=1`, {
         headers: {
             Referer: 'https://tieba.baidu.com/',
         },
     });
     const $ = cheerio.load(data);
-    const max = Number.parseInt($('[max-page]').attr('max-page'));
+    const max = parseInt($('[max-page]').attr('max-page'));
     if (max > pn) {
         return getPost(id, max);
     }
@@ -47,8 +47,8 @@ module.exports = async (ctx) => {
                     const { author, content } = item.data('field');
                     const tempList = item
                         .find('.post-tail-wrap > .tail-info')
-                        .map((_, element) => $(element).text())
-                        .get();
+                        .toArray()
+                        .map((element) => $(element).text());
                     let [pubContent, from, num, time] = ['', '', '', ''];
                     if (0 === tempList.length && 'date' in content) {
                         num = `${content.post_no}楼`;
@@ -68,8 +68,8 @@ module.exports = async (ctx) => {
                             作者：${author.user_name}<br>
                             楼层：${num}<br>
                             ${from}
-                        `, // prettier-ignore
-                        pubDate: new Date(time),
+                        `,
+                        pubDate: timezone(parseDate(time, 'YYYY-MM-DD hh:mm'), +8),
                         link: `https://tieba.baidu.com/p/${id}?pid=${content.post_id}#${content.post_id}`,
                     };
                 })
